refactor(core): split update-worker tracking into status/worker leaf module
Keeps status.rs under the 600-line maintainability target by moving the
worker ownership marker and abandoned-attempt classification into a
focused leaf module, re-exported unchanged.
